refactor: extract shared hook and pass mutation result in SelectEmployees containers
- Extract useSelectEmployeesData hook consolidating employee list fetching,
  mapping, filtering, pagination, selection, and search state shared between
  SelectEmployeesHoliday and SelectEmployeesTimeOff
- Pass result.holidayPayPolicy to onEvent in SelectEmployeesHoliday standalone
  mode, matching the existing pattern in SelectEmployeesTimeOff
- Wrap SelectEmployeesTimeOff handleContinue in useCallback

Addresses code duplication and result-passing feedback from PR #1616 review.
